Copying an entire section and variables approach.
The problem about changing a name of a section in place is that there
is a chance that malloc would return NULL at any step. So for example if
a malloc fails in the section variable moving process then the hash
table gets corrupted.

Copying the section and the variables advantages would be that if malloc
fails then at least I can simply delete the new section and do nothing
with the old section and its variables.